+++ dsa/assets/validate_brackets_CC13.png
