+++ dsa/md_files/linked_list_CC06.md
